+++ SRC/shgeqz.f
@@ -403,7 +403,7 @@
 *     Check Argument Values
 *
       INFO = 0
-      WORK( 1 ) = MAX( 1, N )
+      WORK( 1 ) = REAL( MAX( 1, N ) )
       LQUERY = ( LWORK.EQ.-1 )
       IF( ISCHUR.EQ.0 ) THEN
          INFO = -1
